# webapp/server/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CartRover",
    "description": "API for managing users, products and Cart",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local server"
    }
  ],
  "paths": {
    "/api/v1/users/register": {
      "post": {
        "summary": "Register a new user",
        "description": "Endpoint to register a new user.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User registered successfully"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/v1/users/login": {
      "post": {
        "summary": "User login",
        "description": "Endpoint for user login.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/v1/products/add": {
      "post": {
        "summary": "Add a new product",
        "description": "Endpoint to add a new product.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number"
                  },
                  "images": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "description": {
                    "type": "string"
                  },
                  "category": {
                    "type": "string"
                  }
                },
                "required": ["name", "price", "description", "category"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product added successfully"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/v1/products/": {
      "get": {
        "summary": "Get all products",
        "description": "Endpoint to retrieve all products.",
        "responses": {
          "200": {
            "description": "List of products"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/v1/products/{id}": {
      "get": {
        "summary": "Get a product by ID",
        "description": "Endpoint to retrieve a product by its ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the product to retrieve",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product details"
          },
          "404": {
            "description": "Product not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/v1/cart/add": {
      "post": {
        "summary": "Add a product to the cart",
        "description": "Adds a product to a user's cart.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "user": {
                    "type": "string",
                    "description": "User ID"
                  },
                  "product": {
                    "type": "string",
                    "description": "Product ID"
                  }
                },
                "required": ["user", "product"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product added to cart"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/v1/cart/remove": {
      "post": {
        "summary": "Remove a product from the cart",
        "description": "Removes a product from a user's cart.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "user": {
                    "type": "string",
                    "description": "User ID"
                  },
                  "product": {
                    "type": "string",
                    "description": "Product ID"
                  }
                },
                "required": ["user", "product"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product removed from cart"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    }
  }
}